refactor: extract ui scrollbar helpers
